+++ operator/internal/validation/recordingrule_test.go
@@ -73,7 +73,7 @@ var rtt = []struct {
 			"testing-rule",
 			field.ErrorList{
 				field.Invalid(
-					field.NewPath("Spec").Child("Groups").Index(1).Child("Name"),
+					field.NewPath("spec").Child("groups").Index(1).Child("name"),
 					"first",
 					lokiv1.ErrGroupNamesNotUnique.Error(),
 				),
@@ -95,7 +95,7 @@ var rtt = []struct {
 			"testing-rule",
 			field.ErrorList{
 				field.Invalid(
-					field.NewPath("Spec").Child("Groups").Index(0).Child("Interval"),
+					field.NewPath("spec").Child("groups").Index(0).Child("interval"),
 					"1mo",
 					lokiv1.ErrParseEvaluationInterval.Error(),
 				),
@@ -123,7 +123,7 @@ var rtt = []struct {
 			"testing-rule",
 			field.ErrorList{
 				field.Invalid(
-					field.NewPath("Spec").Child("Groups").Index(0).Child("Rules").Index(0).Child("Record"),
+					field.NewPath("spec").Child("groups").Index(0).Child("rules").Index(0).Child("record"),
 					"invalid&metric:name",
 					lokiv1.ErrInvalidRecordMetricName.Error(),
 				),
@@ -150,7 +150,7 @@ var rtt = []struct {
 			"testing-rule",
 			field.ErrorList{
 				field.Invalid(
-					field.NewPath("Spec").Child("Groups").Index(0).Child("Rules").Index(0).Child("Expr"),
+					field.NewPath("spec").Child("groups").Index(0).Child("rules").Index(0).Child("expr"),
 					"this is not a valid expression",
 					lokiv1.ErrParseLogQLExpression.Error(),
 				),
@@ -177,7 +177,7 @@ var rtt = []struct {
 			"testing-rule",
 			field.ErrorList{
 				field.Invalid(
-					field.NewPath("Spec").Child("Groups").Index(0).Child("Rules").Index(0).Child("Expr"),
+					field.NewPath("spec").Child("groups").Index(0).Child("rules").Index(0).Child("expr"),
 					`{message=~".+"}`,
 					lokiv1.ErrParseLogQLNotSample.Error(),
 				),